+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [2.1.1] - 2020-08-26
+
 ### Changed
 
 - Delete chart-operator release if it stuck in `pending-install` status.
@@ -139,7 +141,8 @@ from Helm 2 to Helm 3.
 
 - Flattening operator release structure.
 
-[Unreleased]: https://github.com/giantswarm/app-operator/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app-operator/compare/v2.1.1...HEAD
+[2.1.1]: https://github.com/giantswarm/app-operator/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/giantswarm/app-operator/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/giantswarm/app-operator/compare/v1.1.11...v2.0.0
 [1.1.11]: https://github.com/giantswarm/app-operator/compare/v1.1.10...v1.1.11